pyproject.toml
[project]
name = "scraping-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "lxml>=6.1.0",
    "pillow>=12.2.0",
    "playwright>=1.59.0",
    "psutil>=7.2.2",
    "pydantic-settings>=2.14.0",
    "pytesseract>=0.3.13",
    "python-json-logger>=4.1.0",
    "python-multipart>=0.0.27",
    "uvicorn>=0.46.0",
]
